Seperate implementation to source file
Moved teamToString to cpp file
